+++ frontend/src/features/products/hooks/useProductActions.ts
@@ -19,8 +19,6 @@ import { ProductWithPopulate } from '../../../types'
 const UseProductActions = ( fetchOnDelete: boolean ) => {
   const dispatch = useAppDispatch()
   const [open, setOpen] = useState(false)
-  const [confirmationOpen, setConfirmationOpen] = useState(false)
-  const [productToDeleteId, setProductToDeleteId] = useState<string | null>(null)
   const products = useAppSelector(selectProductsWithPopulate)
   const [selectedProduct, setSelectedProduct] = useState<ProductWithPopulate | null>(null)
   const { id } = useParams()
@@ -63,26 +61,20 @@ const UseProductActions = ( fetchOnDelete: boolean ) => {
 
   const deleteOneProduct = async (id: string) => {
     try {
-      await dispatch(deleteProduct(id))
-      if (fetchOnDelete) {
-        fetchAllProducts()
+      if (confirm('Вы уверены, что хотите удалить этот товар?')) {
+        await dispatch(deleteProduct(id))
+        if (fetchOnDelete) {
+          fetchAllProducts()
+        } else {
+          navigate('/products')
+        }
+        toast.success('Товар успешно удален.')
       } else {
-        navigate('/products')
+        toast.info('Вы отменили удаление товара.')
       }
-      toast.success('Товар успешно удалён!')
     } catch (e) {
       console.error(e)
     }
-  }
-
-  const handleConfirmationOpen = (id: string) => {
-    setProductToDeleteId(id)
-    setConfirmationOpen(true)
-  }
-
-  const handleConfirmationClose = () => {
-    setConfirmationOpen(false)
-    setProductToDeleteId(null)
   }
 
   return  {
@@ -98,12 +90,7 @@ const UseProductActions = ( fetchOnDelete: boolean ) => {
     id,
     navigate,
     loading,
-    error,
-    confirmationOpen,
-    handleConfirmationOpen,
-    handleConfirmationClose,
-    productToDeleteId,
-  }
+    error }
 }
 
 export default UseProductActions
